Expendedor actualizado usando excepciones

+++ codigo/ComparaPrecio.java
@@ -1,5 +1,5 @@
 public class ComparaPrecio {
-    public static void compara(int pagado, int precio) throws PagoInsuficienteException{
+    public static void compara(int pagado, double precio) throws PagoInsuficienteException{
         if (pagado < precio){
             throw new PagoInsuficienteException();
         }
